Add recent parameter to periods
Sometimes, when requesting someone's periodic data, the start boundary
does not have any data, so it defaults to 0; and returns the stats of
the profile up to the end period.

Using the recent parameter in the query (&recent=true), when this
happens, the profile service will look for the "nearest" most recent
date.